breaking: make with_lease macro, otel module and instance_utils module private
